requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==2.0.0
astor==0.8.1
async-timeout==4.0.2
boto==2.49.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
contextlib2==21.6.0
contourpy==1.0.5
crcmod==1.7
cryptography==37.0.4
cycler==0.11.0
decorator==5.1.1
execnet==1.9.0
fasteners==0.18
fonttools==4.25.0
frozenlist==1.3.3
gcs-oauth2-boto-plugin==3.0
google-apitools==0.5.32
google-auth==2.16.0
google-reauth==0.1.1
gsutil==5.20
h5py==3.7.0
httplib2==0.20.4
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
MedPy==0.4.0
monotonic==1.6
multidict==6.0.4
munkres==1.1.4
nibabel==5.0.1
numpy==1.23.2
oauth2client==4.1.3
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==22.0
paddle-bfloat==0.1.7
paddlepaddle-gpu==2.4.1.post117
path==16.6.0
path.py==12.5.0
Pillow==9.2.0
pip==22.3.1
ply==3.11
protobuf==3.20.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
pytest-shutil==1.7.0
python-dateutil==2.8.2
pyu2f==0.1.5
requests==2.28.2
retry-decorator==1.1.1
rsa==4.7.2
setuptools==65.6.3
SimpleITK==2.2.1
sip==6.6.2
six==1.16.0
termcolor==2.2.0
toml==0.10.2
tornado==6.2
urllib3==1.26.14
wheel==0.38.4
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.8.2
